requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aistudio-sdk==0.3.8
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
av==16.0.1
bce-python-sdk==0.9.57
beautifulsoup4==4.14.3
boto3==1.42.21
botocore==1.42.21
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
datasets==4.4.2
dill==0.4.0
distro==1.9.0
doclayout-yolo==0.0.4
fast-langdetect==0.2.5
fastapi==0.128.0
fasttext-predict==0.9.2.4
ffmpy==1.0.0
filelock==3.20.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
gradio==5.49.1
gradio-client==1.13.3
gradio-pdf==0.0.22
greenlet==3.3.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-retries==0.4.5
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imageio==2.37.2
imagesize==1.4.1
jieba==0.42.1
jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
json-repair==0.55.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain-core==1.2.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
lazy-loader==0.4
loguru==0.7.3
magika==0.6.2
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mineru==2.7.1
mineru-vl-utils==0.1.19.1
mlx==0.30.1
mlx-lm==0.29.1
mlx-metal==0.30.1
mlx-vlm==0.3.9
modelscope==1.33.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.6.1
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.14.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.12
pandas==2.3.3
pdfminer-six==20260107
pdftext==0.6.3
pillow==11.3.0
playwright==1.57.0
pypandoc-binary==1.17
polars==1.36.1
polars-runtime-32==1.36.1
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic-core==2.33.2
pydantic-settings==2.12.0
pydub==0.25.1
pyee==13.0.0
pygments==2.19.2
pyparsing==3.3.1
pypdf==6.5.0
pypdfium2==4.30.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.2
qwen-vl-utils==0.0.14
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
robust-downloader==0.0.2
ruamel-yaml==0.19.1
ruff==0.14.10
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.16.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==79.0.1
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.1
starlette==0.50.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.2
thop==0.1.1.post2209072238
tifffile==2025.12.20
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.3
typer==0.21.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.11.0
ultralytics==8.3.247
ultralytics-thop==2.0.18
urllib3==2.6.2
uuid-utils==0.13.0
uvicorn==0.40.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
